+++ python/py_gdb_thread.c
@@ -292,9 +292,9 @@ sr_py_gdb_thread_str(PyObject *self)
 {
     struct sr_py_gdb_thread *this = (struct sr_py_gdb_thread *)self;
     struct sr_strbuf *buf = sr_strbuf_new();
-    sr_strbuf_append_strf(buf, "Thread #%u with %d frames",
+    sr_strbuf_append_strf(buf, "Thread #%u with %zd frames",
                            this->thread->number,
-                           PyList_Size(this->frames));
+                           (ssize_t)(PyList_Size(this->frames)));
     char *str = sr_strbuf_free_nobuf(buf);
     PyObject *result = Py_BuildValue("s", str);
     free(str);
